+++ src/NetArchTest.Rules/Extensions/TypeDefinitionExtensions.cs
@@ -5,17 +5,8 @@
     using System.Collections.Generic;
     using Mono.Cecil;
 
-    /// <summary>
-    /// Extensions for the <see cref="TypeDefinition"/> class.
-    /// </summary>
-    static internal class TypeDefinitionExtensions
+    internal static class TypeDefinitionExtensions
     {
-        /// <summary>
-        /// Tests whether one class inherits from another.
-        /// </summary>
-        /// <param name="child">The class that is inheriting from the parent.</param>
-        /// <param name="parent">The parent that is inherited.</param>
-        /// <returns>An indication of whether the child inherits from the parent.</returns>
         public static bool IsSubclassOf(this TypeDefinition child, TypeDefinition parent)
         {
             if (parent != null)
@@ -30,11 +21,6 @@
             }
         }
 
-        /// <summary>
-        /// Enumerate the base classes throughout the chain of inheritence.
-        /// </summary>
-        /// <param name="classType">The class to enumerate.</param>
-        /// <returns>The enumeration of base classes.</returns>
         private static IEnumerable<TypeDefinition> EnumerateBaseClasses(this TypeDefinition classType)
         {
             for (var typeDefinition = classType; typeDefinition != null; typeDefinition = typeDefinition.BaseType?.Resolve())
@@ -43,11 +29,6 @@
             }
         }
 
-        /// <summary>
-        /// Convert the definition to a <see cref="Type"/> object instance.
-        /// </summary>
-        /// <param name="typeDefinition">The type definition to convert.</param>
-        /// <returns>The equivalent <see cref="Type"/> object instance.</returns>
         public static Type ToType(this TypeDefinition typeDefinition)
         {
             // Nested types have a forward slash that should be replaced with "+"
