+++ src/NetArchTest.Rules/Extensions/TypeExtensions.cs
@@ -5,30 +5,18 @@
     using System.Reflection;
     using Mono.Cecil;
 
-    /// <summary>
-    /// Extensions for the <see cref="Type"/> class.
-    /// </summary>
-    static internal class TypeExtensions
+    internal static class TypeExtensions
     {
-        /// <summary>
-        /// Converts the value to a <see cref="TypeDefinition"/> instance.
-        /// </summary>
-        /// <param name="type">The type to convert.</param>
-        /// <returns>The converted value.</returns>
         public static TypeDefinition ToTypeDefinition(this Type type)
         {
-            // Get the assembly using reflection
             var assembly = Assembly.GetAssembly(type);
-
-            // Load the assembly into the Mono.Cecil library
-            UriBuilder uri = new UriBuilder(assembly.CodeBase);
-            string path = Uri.UnescapeDataString(uri.Path);
+            var uri = new UriBuilder(assembly.CodeBase);
+            var path = Uri.UnescapeDataString(uri.Path);
             var assemblyDef = AssemblyDefinition.ReadAssembly(path);
 
-            // Find the matching type
             var dependencies = (assemblyDef.Modules
                 .SelectMany(t => t.Types)
-                .Where(t => t.IsClass && t.Namespace != null && t.FullName.Equals(type.FullName, StringComparison.InvariantCultureIgnoreCase)));
+                .Where(t => t.IsClass && t.Namespace != null && t.FullName.Equals(type.FullName)));
 
             return dependencies.FirstOrDefault();
         }
