+++ src/NetArchTest.Rules/Matches/Matcher.cs
@@ -1,8 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
 using Mono.Cecil;
-using static NetArchTest.Rules.FunctionalExtensions;
+using static NetArchTest.Rules.Utils.FunctionalExtensions;
 
 namespace NetArchTest.Rules.Matches
 {
@@ -38,37 +37,6 @@ namespace NetArchTest.Rules.Matches
         public static implicit operator Filter(Matcher matcher)
         {
             return new Filter(input => input.Where(matcher._matchFunc));
-        }
-
-    }
-    
-    public class Filter 
-    {
-        private readonly Func<IEnumerable<TypeDefinition>, IEnumerable<TypeDefinition>> _filterFunc;
-
-        public Filter(Func<IEnumerable<TypeDefinition>, IEnumerable<TypeDefinition>> filterFunc)
-        {
-            _filterFunc = filterFunc;
-        }
-        
-        public static implicit operator Func<IEnumerable<TypeDefinition>, IEnumerable<TypeDefinition>>(Filter filter)
-        {
-            return filter._filterFunc;
-        }
-        
-        public static Filter operator &(Filter left, Filter right)
-        {
-            return new Filter(input => right._filterFunc(left._filterFunc(input)));
-        }
-        
-        public static Filter operator |(Filter left, Filter right)
-        {
-            return new Filter(input => right._filterFunc(input).Concat(left._filterFunc(input)).Distinct());
-        }
-        
-        public static Filter operator !(Filter filter)
-        {
-            return new Filter(input => input.Except(filter._filterFunc(input)));
         }
 
     }
